Push stream mechanism further throughout code and resolve cyclic dependency
We had a cyclic dependency between ElectionService and
MarathonSchedulerService. This has been resolved and the dependency
graph more straightforward by splitting out the initialization logic to
MarathonInitializer.

In order to break the cyclic dependency, some substantial changes to our
DI initialization code had to change. While we could put them back, less
guice is the direction we are going.

We had a small (unlikely, but there nonetheless) race condition in
ElectionServiceImpl subscribe. We subscribe to the event bus, query the
current state, and then send it. It is possible that the state could be
updated between query and send, and the receive would get the events out
of order. By introducing the Repeater component, we remove this side
effect; as an added benefit, we no longer send duplicate events.

ElectionService still needs to know about leader readiness, but, the
back and forth has been substantially reduced. It receives the
MarathonInitializer as a constructor parameter, and applies the
initialize method before publishing the
LeadershipTransition.ElectedAndReady message.

Other components have been updated to use streams instead of the
eventBus, further reducing dependencies between modules. Instead of
depending on the whole electionService, a component can just bring in a
Source[LeadershipTransition, Cancellable]. Some other side benefits are
that if materialized with a Materializer that encloses the context, then
this Source is automatically closed.

Exposing a stream interface for ElectionService will enable other
simplifications, not done in this patch in order to control change
scope.